feat: Implement document monitor
Implemented a document monitor that ensures each committing peer with the 'sidetree' role has all of the document operations that it's supposed to have.
The monitor reads blocks from the ledger looking for batch writes. It traverses through all of the document operations in the batch and ensures that the peer has the operations stored in its DCAS document store.

closes #55

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>