feat: Implement REST API's in the peer process
The external, Sidetree Docker container has been removed and the REST endpoints are now inside of the peer(s). Each peer exposes a REST endpoint according to its role. If the peer has the 'sidetree-resolver' role then the document resolver REST endpoint is exposed. If the peer has the 'sidetree-batch-writer' role then the document update endpoint is exposed.

Also, the name of the 'observer' role was changed to 'sidetree-observer' in order to be more explicit. And the 'sidetree' role was replaced by two roles: 'sidetree-resolver' and 'sidetree-batch-writer'.

closes #66

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>