feat: Store Sidetree configuration on the ledger
Sidetree configuration is now stored on the ledger (which is channel-specific). Multiple Sidetree namespaces may be defined on each channel and a single REST server serves all of the namespaces (on a single port). For example, you can define the following namespace/path combinations:

- Channel: mychannel, Namespace: did:sidtree, Base path: /document
- Channel: mychannel, Namepsace: did:bloc:trustbloc.dev, Base path: /trustbloc.dev
- Channel: yourchannel, Namespace: did:bloc:yourdomain.com, Base path: /yourdomain.com

Dynamic updates are also supported. A new namespace may be added/removed dynamically, or the configuration for an existing namespace may be updated (e.g. a new protocol version added or the base path modified).

closes #65
closes #102

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>