feat: HTTP handlers for files backed by Sidetree
Added a framework to support multiple HTTP endpoints that serve up files (e.g. schemas, .well-known, etc.) which are stored in DCAS.

A file is uploaded to DCAS through one of the configured endpoints and a DCAS ID is returned. The ID is then mapped to a resource name and stored in a Sidetree 'index' document. The ID of the index document is referenced in the ledger configuration for the given HTTP path so that a client may retrieve the file (which is stored in DCAS) using a friendly name.

closes #145

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>