feat: Change structure of file index document
The format of the file index document was changed such that it has explicit fields (as opposed to just a map of key-value pairs). Also, a document validator was added to validate create and update operations of the file index document.

This change required that the file-handler BDD test be reverted back to using file-handler steps and not the fabric-cli, since using the fabric-cli causes a circular dependency (i.e. fabric-cli depends on a sidetree-fabric image for BDD tests).

closes #165

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>