chore: Update to latest fabric-peer-ext
The in-process user chaincode function, GetDBArtifacts, now asks for DB artifacts for specific collections. This allows the document chaincode to be collection-agnostic, i.e. the names of the collections and the number of collections are specified when the chaincode is instantiated.

This chainge also renames the chaincode FilesCC to FileCC.

closes #186

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>