feat: Added blockchain/configblock REST endpoint
Added a REST endpoint that retrieves a config block. If a hash is specified then the block for that hash is retrieved and the last config index is extracted from the metadata section of that block. This index is used to retrieve the config block.
If no hash is specified then the latest block in the chain is retrieved and the config block is retrieved using the last config index of that block.
By default the data sections (data and metadata) of the returned config block are encoded in JSON format. But if the data-encoding parameter is set to 'base64' or 'base64url' then the data sections are encoded in base64 or base64url respectively.

closes #220

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>