feat: Added bearer token authorization handler
Authorization tokens may be configured at the peer config (core.yaml) and then referenced in the REST handler configuration. If tokens are defined for a REST endpoint then an authorization handler is first invoked in order to validate that a valid token was passed in the request header. If the token is valid then the request is allowed to proceed, otherwise an HTTP 401 (UNAUTHORIZED) is returned.

closes #152

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>