feat: Combine Monitor with Observer
The Monitor functionality is added to the Observer so that a separate Monitor is no longer necessary. The Observer now processes batches from block events (as before) and also runs periodically in order to ensure that all blocks are processed. Should processing fail due to a transient error then the block is retried at the next scheduled interval.

closes #312

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>