feat: Added standby observer role
An Observer is active if it has the sidetree-observer role and standby if it has the sidetree-observer-standby role. A standby Observer becomes active when there are no 'alive' active Observers in the cluster.
There may be multiple active Observers and multiple standby Observers, although, only one Observer can process any given block. The metadata has a new field, LeaseOwner, which specifies the endpoint of the peer that currently holds a lease to process a range of blocks. The lease is relinquished after all blocks (up to the current block height) have been processed, at which time a new lease owner is chosen for the next range of blocks (using a deterministic algorithm based on block number).

closes #313

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>