fix: Validate transaction info during 'writeAnchor'
This commit doesn't solve issue #428 as it is written since (in hindsight) this scenario cannot reliably be prevented. If a peer that's behind in protocol version submits an anchor to the blockchain then operations in the anchor can still be processed by peers (Observer) at a higher protocol version since those peers should still support previous protocol versions. So, this commit simply adds validation logic on the endorsers to ensure that the protocol version in the request is valid. If the endorser is behind in protocol version (i.e. behind the version in the request) then the endorser will return an error and the batch writer will retry the request on another peer or until the peer has caught up.

closes #428

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>