feat: Modified Sidetree version endpoint to return protocol version and parameters
The existing Sidetree version endpoint was modified to return the protocol version instead of the component version. The protocol parameters are also returned in the response.

The version endpoint accepts a 'time' parameter which, when specified, returns the protocol version applicable to that transaction time (block number). If not specified then the protocol for the latest block number is returned.

closes #432

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>