chore: Change file handler path to be consistent with DID
The file handler paths for retrieve and update handlers are now prefixed with /identifiers and /operations respectively. Previously there was no prefix and therefore the /version endpoint was being ignored and the /version request was directed to the retrieve handler. This also makes the file endpoint consistent with the DID endpoints.

closes #433

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>