chore: Updated to latest fabric-peer-ext which supports IPFS-based DCAS
Updated to the latest version of fabric-peer-ext which has an IPFS-based DCAS implementation. The DCAS now stores data in standard IPFS format and returns IPFS content IDs (CID).

Files are uploaded to DCAS using a 'file' object format which creates a Merkle DAG for large files.

This change also updates the document store to use an off-ledger client instead of a DCAS client since document operations are not stored in a DCAS collection but an off-ledger collection type (also, the DCAS client no longer supports queries).

closes #459

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>