fix: Fix race condition in document cache
This commit fixes a race condition which exists in the document cache. The scenario is as follows:

- Resolving peer invalidates a document
- Resolve is called before the document is updated in the store by the Observer
- The old document is cached
- Observer stores the new document but the old document is already cached

In order to fix this race condition, instead of deleting the document from the cache, the document is marked stale. So, the next time it is retrieved from the cache it will show as stale and the document will be loaded from the store.

closes #477

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>